requirements.txt
aiogram==2.21
asyncio==3.4.3
Django==4.0.6
djangorestframework==3.13.1
google-api-core==2.8.2
google-api-python-client==2.54.0
google-auth==2.9.1
google-auth-httplib2==0.1.0
httplib2==0.20.4
oauth2client==4.1.3
psycopg2-binary==2.9.3
requests==2.28.1